Simplification and explicit use of brackets for some code blocks

+++ src/CSF.h
@@ -69,7 +69,7 @@ class CSF
     // set point cloud from a one-dimensional array. it defines a N*3 point cloud
     // by the given rows. it is the method used to set point cloud from matlab and
     // numpy
-    void setPointCloud(const double* points, const int rows);
+    void setPointCloud(const double* points, const size_t rows);
 
     // PointCloud set pointcloud
     void setPointCloud(const csf::PointCloud& pc);
