+++ src/point_cloud.cpp
@@ -32,7 +32,7 @@ void csf::PointCloud::computeBoundingBox(Point& bbMin, Point& bbMax) const
     {  // zwm
         const csf::Point& P = at(i);
 
-        for (int d = 0; d < 3; ++d)
+        for (size_t d = 0; d < 3; ++d)
         {
             if (P.u[d] < bbMin.u[d]) { bbMin.u[d] = P.u[d]; }
             else if (P.u[d] > bbMax.u[d]) { bbMax.u[d] = P.u[d]; }
